pyproject.toml
[tool.poetry]
name = "cnn-regression"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pillow = "10.0.1"
matplotlib = "3.8.0"
tqdm = "^4.66.1"
numpy = "1.24.4"
pandas = "1.3.5"
opencv-python-headless = "^4.8.0.76"
torch = "2.0.0"
torchvision = "^0.15.0"
dvc = "^3.30.3"
dvc-gdrive = "^2.20.0"

[tool.poetry.dev-dependencies]
pre-commit = "3.4.0"

[tool.black]
line-length = 90
target-version = ["py310"]
skip-string-normalization = true

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 90
lines_after_imports = 2

#[[tool.poetry.source]]
#name = "torch"
#url = "https://download.pytorch.org/whl/cpu"
#priority = "primary"


#[[tool.poetry.source]]
#name = "torchvision"
#url = "https://download.pytorch.org/whl/cpu"
#priority = "primary"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
